Added Sampler2D class

+++ RayTracer/include/core/utility.h
@@ -11,8 +11,8 @@ inline glm::vec3 clamp(glm::vec3 v)
 	return glm::min(glm::vec3(1.f), glm::max(glm::vec3(0.f), v));
 }
 
-inline void crop(float min, float max, int x, int cropped[])
+inline void crop(float min, float max, size_t x, size_t cropped[])
 {
-	cropped[0] = (int)round(clamp(min) * x);
-	cropped[1] = (int)round(clamp(max) * x);
+	cropped[0] = int(round(clamp(min) * x));
+	cropped[1] = int(round(clamp(max) * x));
 }
